Download data and read into R

These functions are downloading and reading data sets from my courses into an R session. To keep the size of the package manageable, these data sets are not included with the package and must be downloaded separately using download_data(); this only needs to be done once.

Usage

download_data(name)

read_data(name)

Arguments

name

Name of data set to read/download

Examples

download_data('anorexia')   # Download a specific data set
dat <- read_data(whickham)
head(dat)
#>    Smoking    Age Survival
#>     <char> <char>   <char>
#> 1:  Smoker  18-54     Dead
#> 2:  Smoker  18-54     Dead
#> 3:  Smoker  18-54     Dead
#> 4:  Smoker  18-54     Dead
#> 5:  Smoker  18-54     Dead
#> 6:  Smoker  18-54     Dead